Rename pipeline files
