+++ tsdat/pipeline/base.py
